codebase-analyser/requirements.txt
# Core dependencies
tree-sitter==0.20.1
lancedb>=0.4.0  # Updated from 0.3.3 to resolve conflict
numpy==1.24.3
pydantic==2.4.2
networkx==3.1
pandas==2.2.3

# Language-specific grammars
tree-sitter-languages==1.5.0

# For embedding generation
transformers==4.34.0
torch==2.0.1
sentence-transformers==2.2.2
scikit-learn==1.3.0  # For PCA and t-SNE visualization

# Visualization
matplotlib==3.7.2
plotly==5.18.0

# Database
pyarrow==14.0.1

# NLP dependencies
spacy>=3.7.0
# Download the model with: python -m spacy download en_core_web_md

# LangGraph dependencies
langgraph>=0.3.25
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0

# Utilities
pydantic>=2.0.0
tqdm==4.66.1
requests==2.31.0
